# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: hollr.chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels/{channelId}/messages:
    get:
      tags: [Messages]
      operationId: listMessages
      summary: List messages in a channel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - name: before
          in: query
          required: false
          description: Cursor for pagination — fetch messages before this message ID
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Messages]
      operationId: sendMessage
      summary: Send a message to a channel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /channels/{channelId}/messages/{messageId}:
    patch:
      tags: [Messages]
      operationId: editMessage
      summary: Edit a message
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/MessageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
      responses:
        '200':
          description: Message edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Messages]
      operationId: deleteMessage
      summary: Delete a message
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/MessageId'
      responses:
        '200':
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /channels/{channelId}/messages/search:
    get:
      tags: [Messages]
      operationId: searchMessages
      summary: Search messages in a channel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Matching messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'

  # Direct Messages
...
components:
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: string
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # Auth schemas
...
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
        status:
          type: string
          enum: [online, idle, dnd, offline]
          default: offline
        customStatus:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    Message:
      type: object
      required: [id, content, authorId, createdAt, author]
      properties:
        id:
          type: string
        content:
          type: string
        authorId:
          type: string
        channelId:
          type: [string, 'null']
        dmThreadId:
          type: [string, 'null']
        edited:
          type: boolean
          default: false
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        author:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 0
          maxLength: 4000
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentInput'
    EditMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000

    # Attachment schemas
    Attachment:
      type: object
      required: [id, objectPath, name, contentType, size]
      properties:
        id:
          type: string
        objectPath:
          type: string
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer
    AttachmentInput:
      type: object
      required: [objectPath, name, contentType, size]
      properties:
        objectPath:
          type: string
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer

    # DM schemas
...